[WIP] add remote origin headers also in error case
