+++ app/controllers/Controller.scala
@@ -2,9 +2,8 @@ package controllers
 
 import com.mohiva.play.silhouette.api.actions.{SecuredRequest, UserAwareRequest}
 import com.scalableminds.util.accesscontext.{AuthorizedAccessContext, DBAccessContext}
-import com.scalableminds.util.mvc.ExtendedController
+import com.scalableminds.util.mvc.{ExtendedController, ValidationHelpers}
 import com.scalableminds.util.tools.Fox
-import com.scalableminds.webknossos.datastore.controllers.ValidationHelpers
 import com.typesafe.scalalogging.LazyLogging
 import models.user.User
 import oxalis.security.{UserAwareRequestLogging, WkEnv}
