+++ webknossos-datastore/app/com/scalableminds/webknossos/datastore/controllers/StandaloneDatastore.scala
@@ -2,17 +2,17 @@ package com.scalableminds.webknossos.datastore.controllers
 
 import javax.inject.Inject
 import play.api.libs.json.Json
-import play.api.mvc.{Action, AnyContent, InjectedController}
+import play.api.mvc.{Action, AnyContent}
 
-class StandaloneDatastore @Inject()() extends InjectedController with RemoteOriginHelpers {
+class StandaloneDatastore @Inject()() extends Controller {
+
+  override def allowRemoteOrigin: Boolean = true
 
   def buildInfo: Action[AnyContent] = Action {
-    AllowRemoteOrigin {
-      Ok(
+    Ok(
         Json.obj(
           "webknossosDatastore" -> webknossosDatastore.BuildInfo.toMap.mapValues(_.toString),
           "webknossos-wrap" -> webknossoswrap.BuildInfo.toMap.mapValues(_.toString)
         ))
-    }
   }
 }
